installed react-router-dom and added Contacts page
i also added css code to hide the scrollbar. it work on both chrome and edge, but not in firefox 89.0.2. In this commit i create a route for different pages that consist on my website. the route is defined in App.js. I also cleaned up some css style so it looks cleaner now.