fixed font-size in mobile screen
before fixed, the style that used to change font-size when the vnav lis are active are inline. because of that, i can't insert it inside media query to adapt mobile screen . i changed this inline style with classlist.add(), whenever active state is change. the function that contains that is then called by useLayoutEffect (handleActive)